requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pdf2image>=1.16.3
pytesseract>=0.3.10
scikit-image>=0.21.0
matplotlib>=3.7.0
tqdm>=4.65.0 